[unrated] Title: 점 찍기, Time: 0.01 ms, Memory: 4.13 MB -BaekjoonHub

+++ 프로그래머스/unrated/140107. 점 찍기/점 찍기.cpp
@@ -8,8 +8,8 @@ long long solution(int k, int d) {
     long long answer = 0;
     long long dd = (long long)d * d;
     for(int i = 0; i <= d; i += k){
-        long long y = sqrt(dd - (long long)i*i);
-        answer += y/(long long)k + 1;
+        int y = sqrt(dd - (long long)i*i);
+        answer += y/k + 1;
     }
     return answer;
 }
